# backend/docs/swagger.yaml
basePath: /api/v1/
definitions:
  github_com_0ScPro0_affiliate-system_internal_core_transport_dto.CategoryResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  github_com_0ScPro0_affiliate-system_internal_core_transport_dto.CityResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  github_com_0ScPro0_affiliate-system_internal_core_transport_dto.UpdateCategoryRequest:
    properties:
      description:
        maxLength: 1000
        minLength: 1
        type: string
      id:
        type: integer
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - id
    type: object
  github_com_0ScPro0_affiliate-system_internal_core_transport_dto.UpdateCityRequest:
    properties:
      id:
        type: integer
      name:
        maxLength: 50
        minLength: 1
        type: string
    required:
    - id
    type: object
...
  github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
host: 127.0.0.1:8080
info:
  contact: {}
  description: Affiliate System API
  title: Affiliate System API
  version: "1.0"
paths:
...
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Category deleted successfully
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete category
      tags:
      - categories
    get:
      consumes:
      - application/json
      description: Get category by its unique identifier
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category found
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_dto.CategoryResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
      summary: Get category by ID
      tags:
      - categories
    patch:
      consumes:
      - application/json
      description: Update existing category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: UpdateCategory request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_dto.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Category updated successfully
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_dto.CategoryResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update category
      tags:
      - categories
...
  /cities/{id}:
    delete:
      consumes:
      - application/json
      description: Delete city by ID
      parameters:
      - description: City ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: City deleted successfully
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete city
      tags:
      - cities
    get:
      consumes:
      - application/json
      description: Get city by its unique identifier
      parameters:
      - description: City ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: City found
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_dto.CityResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
      summary: Get city by ID
      tags:
      - cities
    patch:
      consumes:
      - application/json
      description: Update existing city by ID
      parameters:
      - description: City ID
        in: path
        name: id
        required: true
        type: integer
      - description: UpdateCity request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_dto.UpdateCityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: City updated successfully
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_dto.CityResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update city
      tags:
      - cities
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by a space and your access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"